Revamp, made a working prototype!
I basically rewrote the entire thing and implemented most of the systems to really get everything working. I also build a fun little tester level. Controls are severely unpolished, camera def needs a tweener, but otherwise standard Celeste clone right here.

Oh, controls are Space to jump, O to dash, WASD to move.

It sucks, but I need something to be able to work with when swapping between testing and tweaking often

Oh you can also draw. No delete specific platform feature yet though, so be careful.

Z to save
X to load
C to clear level
Click to set point 1 or 2 of platform.
1, 2, 3, 0 for different types of platforms

So yeah
Its crappy but it works
Constructive Criticism welcome, but ideally tweaks to features and recommendations for more features are recommended as is the focus of the prototyping phase.